requirements.txt
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
pip~=21.1.2
wheel~=0.36.2
certifi~=2022.6.15
requests~=2.28.1
idna~=3.3
urllib3~=1.26.11
setuptools~=57.0.0
loguru~=0.6.0

telebot~=0.0.5